refactor: Cache auth() function to avoid duplicate calls on the same route
