+++ src/lib/auth.ts
@@ -1,24 +1,14 @@
-import { Role } from "@prisma/client";
+import { cache } from "react";
 
-import { DEFAULT_LOGIN_REDIRECT, instructorRoutePrefix } from "~/routes";
-import { auth } from "~/server/auth";
+import { auth as nextAuth } from "~/server/auth";
+
+/**
+ * Deduplicate the auth() request from the same route
+ */
+export const auth = cache(nextAuth);
 
 export const currentUser = async () => {
   const session = await auth();
 
   return session?.user;
-};
-
-export const getLoginRedirectUrl = ({
-  role = Role.USER,
-}: {
-  role?: Role;
-}): string => {
-  let redirectUrl = DEFAULT_LOGIN_REDIRECT;
-
-  if (role === Role.ADMIN) {
-    redirectUrl = instructorRoutePrefix;
-  }
-
-  return redirectUrl;
 };
